Add savings blurb to book details on mobile
Issue #1891
