+++ src/app/pages/details/desktop-view/details-tab/details-tab.scss
@@ -131,25 +131,4 @@
             top: 0;
         }
     }
-
-    .savings-blurb {
-        align-items: center;
-        display: grid;
-        grid-column-gap: 1rem;
-        grid-template-columns: auto 1fr;
-
-        .book-icon {
-            align-items: center;
-            border-radius: 50%;
-            display: flex;
-            font-size: 2.7rem;
-            height: 6rem;
-            justify-content: center;
-            width: 6rem;
-        }
-
-        span[id] {
-            font-weight: bold;
-        }
-    }
 }
