+++ src/app/pages/details/phone-view/details-pane/details-pane.js
@@ -2,6 +2,7 @@ import React, {useState, useEffect} from 'react';
 import {RawHTML} from '~/components/jsx-helpers/jsx-helpers.jsx';
 import {Authors, PublicationInfo} from '../../common/common';
 import CollapsingPane from '~/components/collapsing-pane/collapsing-pane.jsx';
+import SavingsBlurb from '../../common/savings-blurb';
 import WrappedJsx from '~/controllers/jsx-wrapper';
 import './details-pane.css';
 
@@ -9,6 +10,8 @@ export default function DetailsPane({polish, model}) {
     return (
         <div className="details-pane">
             <RawHTML html={model.description} />
+            <hr className="thin-rule" />
+            {model.adoptions && <SavingsBlurb model={model} />}
             <div class="authors-region">
                 <CollapsingPane title={polish ? 'Autorzy' : 'Authors'}>
                     <Authors model={model} polish={polish} />
