+++ src/app/pages/details/phone-view/details-pane/details-pane.scss
@@ -14,4 +14,13 @@
         margin: 1rem 0;
     }
 
+    .thin-rule {
+        border: 0;
+        border-top: thin solid ui-color(form-border);
+        margin: 2rem 0;
+    }
+
+    .savings-blurb {
+        margin-bottom: 4rem;
+    }
 }
